requirements.txt
Django>=4.2.10,<5
backports.zoneinfo;python_version<"3.9"
pytz==2024.1
resolver==0.2.1
selector==0.10.1
Paste==3.10.0
PasteDeploy==3.1.0
PasteScript==3.5.1
httplib2==0.22.0
feedparser==6.0.8
Markdown==3.6 # pyup: < 3
pyenchant==3.2.2
psycopg2==2.9.9

Pillow==10.3.0

chardet==5.2.0
idna==3.7
urllib3==2.2.1  # sentry-sdk
certifi==2024.2.2  # sentry-sdk

requests==2.31.0
pysolr==3.9.0
requirements/src/gdata-2.0.14.tar.gz
django-tinymce==4.0.0  # pyup: <2.7.0
coverage==7.5.0
pyasn1==0.6.0
pyOpenSSL==24.1.0
ndg-httpsclient==0.5.1
requirements/src/djangohelpers-0.22-py2.py3-none-any.whl
logilab-common==2.0.0
astroid==3.1.0
pylint==3.1.0
pep8==1.7.1
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.2.0
configparser==7.0.0
importlib-metadata<7.2 # for flake8
typing_extensions==4.11.0
zipp==3.18.0  # for flake8
flake8==7.0.0
statsd==4.0.1
six==1.16.0
python-dateutil==2.9.0
text-unidecode==1.3  # for faker
Faker==25.0.0
contextlib2==21.6.0
argparse==1.4.0
cffi==1.16.0
colorama==0.4.4
cryptography==42.0.4
idna==3.7
ipaddress==1.0.23
lazy-object-proxy==1.10.0
nose==1.3.7
pycparser==2.22
wrapt==1.16.0
sqlparse==0.5.0

requirements/src/sorl-3.4-py3-none-any.whl
requirements/src/tagging-0.3-pre.tar.gz
django-indexer==0.3.0
django-templatetag-sugar==1.0
django-statsd-mozilla==0.4.0
django-paging==0.2.5
django-appconf==1.0.5
raven==6.10.0
sentry-sdk==2.0.0
requirements/src/django_googlecharts-1.5-py2.py3-none-any.whl

django-stagingcontext==0.1.0
django-smoketest==1.2.0

django-treebeard==4.7

django-haystack==3.2.1

django-bootstrap-form==3.4
django-bootstrap3==24.2
django-markwhat==1.6.2
django-pagetree==1.5.0
django-pageblocks==1.2.0 # pyup: <2.0 
requirements/src/django_databrowse-2024.3.5-py2.py3-none-any.whl
factory_boy==3.3.0

django-debug-toolbar==4.3.0 # pyup: <2.0
gunicorn==22.0.0

s3transfer==0.10.0
jmespath==1.0.0

boto3>=1.12.0,<1.35.0
botocore>=1.15.0,<1.35.0

urllib3==2.2.1
django-storages==1.14.2
django-impersonate==1.9.1
django-ga-context==0.1.0
django-smtp-ssl==1.0
django-extensions==3.2.0

ctlsettings==0.3.1

pbr==6.0.0
pyyaml==6.0.1
stevedore>=1.20.0 # Apache-2.0
bandit==1.7.1

entrypoints==0.4
distro==1.9.0

python-cas==1.6.0
django-cas-ng==5.0.1

lxml==5.2.0
asgiref==3.8.0
